+++ SIGED/src/indicadores/servlets/CadastrarIndicadorServlet.java
@@ -21,8 +21,12 @@ public class CadastrarIndicadorServlet extends HttpServlet {
 		String nomeIndicador = request.getParameter("nomeIndicador").trim(); // campo obrigatório		
 		String dataIndicador = request.getParameter("dataIndicador").trim(); // campo obrigatório		
 		String depIndicador = request.getParameter("depIndicador").trim(); // campo obrigatório		
-		int metaIndicador = Integer.parseInt(request.getParameter("metaIndicador").trim()); // campo obrigatório
-		int valorIndicador = Integer.parseInt(request.getParameter("valorIndicador").trim()); // campo obrigatório		
+		double metaIndicador = Integer.parseInt(request.getParameter("metaIndicador").trim()); // campo obrigatório
+//				Double.parseDouble(request.getParameter("valorIndicador").trim()); 
+//				Integer.parseInt(request.getParameter("metaIndicador").trim()); // campo obrigatório
+		double valorIndicador = Integer.parseInt(request.getParameter("valorIndicador").trim()); // campo obrigatório
+//				Double.parseDouble(request.getParameter("metaIndicador").trim()); 
+//				Integer.parseInt(request.getParameter("valorIndicador").trim()); // campo obrigatório		
 				
 
 		// Encaminhar para a classe especialista
